#171 look for closest default section to load the UI
