+++ scripts/scripts.js
@@ -443,3 +443,22 @@ export function createIframe(url, { parentEl, classes = [] }) {
 
   return iframe;
 }
+
+/**
+ * search recursively through the parent elements chain until find the target
+ * element based on its tag or a CSS class
+ * @param {HTMLElement} el the initial element or one of its parents
+ * @param {object} option an option object using the key as search parameter
+ * @param {string} option.tag to search by element tag
+ * @param {string} option.className to search by a CSS class in the element
+ * @returns {HTMLElement | null} the target element or null if is not found
+ */
+export const getTargetParentElement = (el, option) => {
+  const [key] = Object.keys(option);
+  const checks = {
+    tag: el.parentElement.localName === option[key],
+    className: el.parentElement.classList.contains(option[key]),
+  };
+  if (checks[key] === undefined) return null;
+  return checks[key] ? el.parentElement : getTargetParentElement(el.parentElement, option);
+};
